pyproject.toml
[tool.poetry]
name = "yolo-from-scratch"
version = "0.1.0"
description = "Model YOLO implementation with testing and best practices"
authors = ["Carlos Salguero"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
numpy = "^1.24.0"
albumentations = "^1.3.1"
pytorch-lightning = "^2.1.0"
hydra-core = "^1.3.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"